finished import process for person, added zp4
git-svn-id: https://svn.cccs.umn.edu/svn/ncs/ncs-etl@264 59f948a9-288f-427a-98b3-c46ff3be0d89

+++ grails-app/domain/edu/umn/ncs/staging/ContactImportLink.groovy
@@ -2,17 +2,17 @@ package edu.umn.ncs.staging
 
 class ContactImportLink {
 	
-	Integer personId
-	Integer primaryPhoneId
-	Integer secondaryPhoneId
-	Integer homePhoneId
-	Integer workPhoneId
-	Integer cellPhoneId
-	Integer emailAddressId
-	Integer addressId
-	Integer dwellingUnitId
-	Integer appointmentId
-	Integer trackedItemId
+	Long personId
+	Long primaryPhoneId
+	Long secondaryPhoneId
+	Long homePhoneId
+	Long workPhoneId
+	Long cellPhoneId
+	Long emailAddressId
+	Long addressId
+	Long dwellingUnitId
+	Long appointmentId
+	Long trackedItemId
 	String norcSuId
 	
 	static belongsTo = [contactImport : ContactImport ]
